requirements.txt
streamlit
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dateutil
tzlocal
sqlalchemy
psycopg
discord.py
langchain
langchain-google-genai
apscheduler
pydantic>=2.0

